+++ app/models/taxon_name_classification/iczn/unavailable/nomen_nudum.rb
@@ -11,7 +11,6 @@ class TaxonNameClassification::Iczn::Unavailable::NomenNudum < TaxonNameClassifi
     'nomen nudum: ' + type_name.demodulize.underscore.humanize.downcase.gsub(/\d+/, ' \0 ').squish
   end
 
-
   def self.disjoint_taxon_name_classes
     self.parent.disjoint_taxon_name_classes +
         self.collect_descendants_and_itself_to_s(TaxonNameClassification::Iczn::Unavailable::Excluded,